+++ app/src/main/java/com/weisj/pj/view/choosearea/WheelView.java
@@ -33,6 +33,7 @@ import android.view.ViewGroup.LayoutParams;
 import android.view.animation.Interpolator;
 import android.widget.LinearLayout;
 
+
 import com.weisj.pj.R;
 import com.weisj.pj.view.ItemsRange;
 
@@ -630,7 +631,7 @@ public class WheelView extends View {
 		/*/ Modified by wulianghuan 2014-11-25
 		int height = (int)(1.5 * getItemHeight());
 		//*/
-		int height = (int)(3 * getItemHeight());
+		int height = 3 * getItemHeight();
 		//*/
 		topShadow.setBounds(0, 0, getWidth(), height);
 		topShadow.draw(canvas);
